requirements.txt
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
python-dateutil==2.9.0.post0
